+++ feature/sponsor/src/main/java/io/github/droidkaigi/confsched2019/sponsor/ui/SponsorFragment.kt
@@ -16,13 +16,13 @@ import dagger.Module
 import dagger.Provides
 import io.github.droidkaigi.confsched2019.di.PageScope
 import io.github.droidkaigi.confsched2019.ext.android.changed
+import io.github.droidkaigi.confsched2019.item.DividerItem
 import io.github.droidkaigi.confsched2019.model.LoadingState
 import io.github.droidkaigi.confsched2019.model.Sponsor
 import io.github.droidkaigi.confsched2019.model.SponsorCategory
 import io.github.droidkaigi.confsched2019.sponsor.R
 import io.github.droidkaigi.confsched2019.sponsor.databinding.FragmentSponsorBinding
 import io.github.droidkaigi.confsched2019.sponsor.ui.actioncreator.SponsorActionCreator
-import io.github.droidkaigi.confsched2019.sponsor.ui.item.DividerItem
 import io.github.droidkaigi.confsched2019.sponsor.ui.item.HeaderItem
 import io.github.droidkaigi.confsched2019.sponsor.ui.item.SponsorItem
 import io.github.droidkaigi.confsched2019.sponsor.ui.item.TallSponsorItem
